Create GCN
At present，the graph convolutional nerve has become a research hotspot.Many problems are graphs in true nature. In our world, we see many data are graphs, such as molecules, social networks, and paper citations networks. In the graph, we have node features (the data of nodes) and the structure of the graph (how nodes are connected).